requirements.txt
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.42.2
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.1.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-filter==24.3
django-modelcluster==6.4
django-permissionedforms==0.1
django-stubs-ext==5.1.3
django-taggit==6.1.0
django-tasks==0.6.1
django-treebeard==4.7.1
djangorestframework==3.15.2
dotenv==0.9.9
draftjs-exporter==5.0.0
et_xmlfile==2.0.0
filetype==1.2.0
gunicorn==23.0.0
idna==3.10
laces==0.1.2
openpyxl==3.1.5
packaging==24.2
pillow==11.1.0
pillow_heif==0.21.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
telepath==0.3.1
typing_extensions==4.12.2
urllib3==2.3.0
wagtail==6.4
wagtail-headless-preview==0.8.0
whitenoise==6.9.0
Willow==1.9.0
